Hide Concurrent instance in Compiler.Target
