+++ core/shared/src/main/scala/fs2/Compiler.scala
@@ -116,17 +116,21 @@ object Compiler extends CompilerLowPriority {
 
   sealed trait Target[F[_]] extends Resource.Bracket[F] {
     def ref[A](a: A): F[Ref[F, A]]
-    def concurrent: Option[Concurrent[F]]
+    def uncancelable[A](f: Poll[F] => F[A]): F[A]
+    private[fs2] def interruptContext(root: Token): Option[F[InterruptContext[F]]]
   }
 
   private[fs2] trait TargetLowPriority {
     implicit def forSync[F[_]: Sync]: Target[F] = new SyncTarget
 
+    private val idPoll: Poll[Id] = new Poll[Id] { def apply[X](fx: Id[X]) = fx }
+
     private final class SyncTarget[F[_]](protected implicit val F: Sync[F])
         extends Target[F]
         with Resource.Bracket.SyncBracket[F] {
       def ref[A](a: A): F[Ref[F, A]] = Ref[F].of(a)
-      def concurrent: Option[Concurrent[F]] = None
+      def uncancelable[A](f: Poll[F] => F[A]): F[A] = f(idPoll.asInstanceOf[Poll[F]])
+      private[fs2] def interruptContext(root: Token): Option[F[InterruptContext[F]]] = None
     }
   }
 
@@ -139,8 +143,10 @@ object Compiler extends CompilerLowPriority {
     ) extends Target[F]
         with Resource.Bracket.MonadCancelBracket[F] {
       def ref[A](a: A): F[Ref[F, A]] = F.ref(a)
-      def uncancelable[A](fa: F[A]): F[A] = F.uncancelable(_ => fa)
-      def concurrent: Option[Concurrent[F]] = Some(F)
+      def uncancelable[A](f: Poll[F] => F[A]): F[A] = F.uncancelable(f)
+      private[fs2] def interruptContext(root: Token): Option[F[InterruptContext[F]]] = Some(
+        InterruptContext(root, F.unit)
+      )
     }
   }
 }
